Add files via upload
The Holliday Junction was constructed using caDNAno (.json), using the M13mp18 sequence (.txt). Using python, the structural data within the .json files was paired with the sequence file information to produce a protein database file (.pdb) for use in all atom simulations. The toplogy and input coordinate files (.prmtop and .inpcrd respectively) are included if you would like to reproduce the simulation or run a longer one if you please. A few clips of a 24 nano-second simulation of the structure are provided (.mov).